requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
comtypes==1.1.8
googlesearch-python==2020.0.2
html5lib==1.1
idna==2.10
pafy==0.5.5
pypiwin32==223
pyttsx3==2.90
pywin32==300
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
urllib3==1.25.9
webencodings==0.5.1
youtube-dl==2021.1.8
